requirements.txt
pywinauto>=0.6.8
python-dotenv>=0.19.0
schedule>=1.1.0
tqdm>=4.65.0 